file struct update & test file update
