--- aimmoPost/config/__init__.py
@@ -1,3 +1,0 @@
-from . import default
-
-default = default
